Move shared ref formatting into AshRefCommand.Formatting inner class
